AI Prompt:
Rename the backend to "server" and the name of the Django app to "RackMaster". Then create a models with the following relationships:

"Racks" with filled with "Unit" where a unit can be a server, switch, NAS or PSU. Each "Unit" in the "Rack" may contain "devices" such as a NIC, SSD, or other such devices. A rack, unit or device will contain a serial number that needs to be logged and an optional status messags(s) with severity of: INFO, WARNING, or CRITICAL. Each "Unit" may also have "connection" to another unit.

Create a devcontainer setup using the existing dockerfiles